Update GAME 6.py
kicking updates after collisions so kicking into walls is fixed
bouncing quotient used in updating positions after collisions as well
added back in 5 players since spawing on top of each other is ok again with new collisions
made it so you can't hold down kick
bugs:
if pink and red push against each other in equal and opposite directions, pink will win
collisions are still not exactly right. push the ball against the corner and kick, then compare to the actual game. Not sure what's causing this
goalpost collision. The idea is that if the ball intersects the goalpost by a certain amount, its position should be updated so that it is that same amount away from the edge of the goalpost, multiplied by the bouncing coefficient. This reflects what was supposed to be happening during the frame while it was intersecting the goalpost. Somehow, the distance variable doesn't represent the distance between the object and the goalpost, although it seems like it should